Construct(add) cb_frame_number_list from list_access
