+++ Makefile.am
@@ -206,7 +206,7 @@ libgts_la_SOURCES = \
 					sources/libcpp83gts_callback_and_action/cb_color_trace_thickness.cpp \
 					sources/libcpp83gts_callback_and_action/cb_color_trace_view_area.cpp \
 					sources/libcpp83gts_callback_and_action/cb_config_save.cpp \
-					sources/libcpp83gts_callback_and_action/cb_fnum_edit_insert.cpp \
+					sources/libcpp83gts_callback_and_action/cb_file_number_list.cpp \
 					sources/libcpp83gts_callback_and_action/cb_quit.cpp \
 					sources/libcpp83gts_callback_and_action/cb_read_and_preview.cpp \
 					sources/libcpp83gts_callback_and_action/cb_read_and_save.cpp \
@@ -230,8 +230,6 @@ libgts_la_SOURCES = \
 					sources/libcpp83gts_callback_and_action/iipg_save.cpp \
 					sources/libcpp83gts_callback_and_action/iipg_scan.cpp \
 					sources/libcpp83gts_callback_and_action/iipg_view.cpp \
-					sources/libcpp83gts_callback_and_action/image_mark_access.cpp \
-					sources/libcpp83gts_callback_and_action/list_access.cpp \
 					sources/libcpp83gts_callback_and_action/memory_config_load.cpp \
 					sources/libcpp83gts_callback_and_action/memory_config_save.cpp \
 					sources/libcpp83gts_callback_and_action/memory_config_trace_src_hsv_load.cpp \
