+++ sources/libcpp81gts_file_browser/gtsfbro06cb_trace_batch_add.cpp
@@ -12,7 +12,7 @@ void gtsfbro06cb_trace_batch::cb_add( void )
 		return; /* 選択されていない --> 正常終了 */
 	}
 
-	/* run listすべてunselectする */
+	/* run listすべて選択状態を外す */
 	for (ii=1; ii<= cl_gts_gui.selbro_trace_batch_run_list->size();
 	++ii) {
 		cl_gts_gui.selbro_trace_batch_run_list->select( ii, 0 );
